Use ADL to retreive generic stream and equal attached to a Tagged Object.
Try to ensure that the resulting struct can still be intialized using aggregate initialization so a constructor is not necessary.

Provide a generic stream operator for structs, provided they implement the 2 required function  refllection functions.

explore some of the feature of C++14 available in VS2015: template type alias and parameter pack.