Add grouped_by file and some basic code to exercise it.
Use it with the by_month() function. Note, because of the use of binary_negate, grouped_by functor does not accept lambda.

There may be a performance issue as creating the range will imediately look for the first result failing the predicate since it uses the next value as the last value of the returned range.

The alternative would be to write a take_while_iterator, that would be somewhat similar to skip_iterator, but woulld just return every elements and skip to last when the predicate fails.

Note:
similar in spirit to v3 group_by but following the transformed, filtered... naming.
uses the internal iterator that is part of adjacent_filtered.